permissiondapp/requirements.txt
pyteal>=0.26.1
python-dotenv>=1.0.1
# development
pytest>=8.3.4
black>=24.10.0